Add files via upload
This game starts  by running welcome_screen.py. from there you have two buttons, only the play is functional. When the play button is clicked, a function containing the game is called, from the file 'game2.py'. When the game is over, a function containing the end screen is called, when main menu is clicked, you're redirected back to the main menu.

The purpose of the game is to click a ball which moves randomly when clicked as many times as possible in 10 seconds. If the countdown reaches 0 or a click outside the ball is detected, the game ends.
welcome_screen => game2 => game_over => welcome_screen.
                             || 
                           stats(no functionality)